Arrumei alguns erros de palavras erradas que estavam atrapalhando o funcionamento das rotas e adicionei um logo provisório
