+++ src/index.html
@@ -6,8 +6,9 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <link rel="stylesheet" href="style.css"/>
     <link rel="stylesheet" href="responsive.css"/>
+    <link rel="shortcut icon" href="img/favicon.png">
 
-    <title>Document</title>
+    <title>Bookish</title>
   </head>
   <body>
     <main id="root"></main>
