+++ src/main.js
@@ -1,34 +1,21 @@
-import feed from './pages/feed/index.js'
-import signup from './pages/signup/index.js';
+import home from './pages/home/index.js'
 import signin from './pages/signin/index.js';
-import createprofile from './pages/createprofile/index.js'
-import profile from './pages/profile/index.js';
-import {loginPage, createUser, asyncSendProfileData,uploadImage, currentUser, getUser, setPersistence} from './lib/index.js'
+import {getUser} from './lib/index.js'
 import { changeContent } from './routes.js';
 //import { router } from './routes.js';
 
 
 const mainLogin = document.getElementById("root")
 
-
-
-
 window.addEventListener("load", () => {
   const loggedUser = getUser()
   console.log(loggedUser)
   if(loggedUser == undefined || loggedUser == null){
     mainLogin.appendChild(signin())
-
-  }else{
-    mainLogin.appendChild(feed())
-    
   }
-  // const url = window.location.pathname
-  // const path = url.slice(1)
-  // console.log(path)
-  // changeContent(path)
- 
-
+  else{
+    mainLogin.appendChild(home())    
+  }
   window.addEventListener('popstate', (e) =>{
     console.log(e.state)
