+++ src/pages/createprofile/index.js
@@ -1,8 +1,6 @@
 import { currentUser , uploadImage, asyncSendProfileData} from "../../lib/index.js"
 import { changeContent} from '../../routes.js'
 import { errorInput } from '../../error.js'
-import feed from "../feed/index.js"
-
 
 export default () => {
 
@@ -15,6 +13,7 @@ export default () => {
 
   const createProfileTemplate = `
     <div class="logo-container logo-container-register">
+    <h1 id="bookish">BOOKISH</h1>
       <img class="site-logo" src="./img/logo.png">
     </div>
     <div  class="form-container" id="form-create-profile">
@@ -91,7 +90,7 @@ export default () => {
     
         })
         .then(()=>{
-          changeContent("feed")
+          changeContent("home")
         })  
       }
     }
